src/tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
aiohttp==3.9.1
cryptography==41.0.7
pyyaml==6.0.1
prometheus-client==0.19.0
